fix: refresh token issues (#396)
* Don't log out upon a network error

* Remove unused component

* Chain refresh -> logout -> clear storage

This should ensure that users are never stuck in a state where they are
logged-in but unable to be properly authenticated (the site will
fail-fast in terms of logging them out). However, this currently means
that users will be logged out upon a network error (this will be fixed
in a future commit).