Handle reload-prompt properly
Now that we stay on the editor page sometimes after submission, we need
to fix the handling of the reload prompt (otherwise, it will never
display after the form is submitted for the first time). To do this, we
check if the form state has been touched (dirtied), and only allow
reloading if it has not. We also have to reset the form default values
as this is how `rhf` determines whether the form is dirty.