feat: better assessment editor loading state (#585)
* Add full-page whiteout instead of loading spinner

Also don't redirect away from the assessment-editor after saving.

* Handle reload-prompt properly

Now that we stay on the editor page sometimes after submission, we need
to fix the handling of the reload prompt (otherwise, it will never
display after the form is submitted for the first time). To do this, we
check if the form state has been touched (dirtied), and only allow
reloading if it has not. We also have to reset the form default values
as this is how `rhf` determines whether the form is dirty.

* PR feedback: don't show error toast on in-editor form validation error

* PR feedback: fix reload-prompt behaviour

* fix: wonky behaviour on create

* fix: stay on page after create

* fix: support redirect on publish-create

* fix: extra notification appearing upon save

* PR feedback: handleCloseEditor

* PR feedback: nits